Three more tests and move csv tests to their own file
